Compare fits

Compare empirical distribution of counts with the distribution defined by the model fitted to counts.

compare_fit(count_list, fitlist = fit_counts(count_list, model = "all"))

Arguments

count_list
A list of counts. Each count should be in separate column, rows should represent values of these counts.

fitlist
a list of fits, as created by fit_counts.

Value

A data.frame with distribution values for each unique count. Count is the name of the original count, model is the name of distribution model, x is unique count value, n is the frequency of unique counts, value is result of calculations made by chosen distribution model.

Examples

df <- data.frame(poisson = rpois(25, 0.3), binomial = rbinom(25, 1, 0.8))
compare_fit(df, fitlist = fit_counts(df, model = "all"))

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#>       count   model x  n      value
#> 1   poisson Poisson 0 20 18.8945935
#> 2   poisson Poisson 1  3  5.2904862
#> 3   poisson Poisson 2  2  0.7406681
#> 4   poisson     ZIP 0 20 20.0000001
#> 5   poisson     ZIP 1  3  3.4229218
#> 6   poisson     ZIP 2  2  1.2244058
#> 7   poisson      NB 0 20 19.8595248
#> 8   poisson      NB 1  3  3.8036842
#> 9   poisson      NB 2  2  0.9651766
#> 10  poisson    ZINB 0 20 19.9999902
#> 11  poisson    ZINB 1  3  3.4229376
#> 12  poisson    ZINB 2  2  1.2243884
#> 13 binomial Poisson 0  3 10.3695728
#> 14 binomial Poisson 1 22  9.1252241
#> 15 binomial     ZIP 0  3 10.3696140
#> 16 binomial     ZIP 1 22  9.1251961
#> 17 binomial      NB 0  3 10.3696041
#> 18 binomial      NB 1 22  9.1251890
#> 19 binomial    ZINB 0  3 10.3695956
#> 20 binomial    ZINB 1 22  9.1251863