Fit counts to distributions

Fit counts to distributions

fit_counts(counts_list, separate = TRUE, model, level = 0.95, ...)

Arguments

counts_list
A list of count data. Each count should be in separate column, rows should represent values of that counts.

separate
logical. If TRUE, each count is separately fitted to the model. If FALSE, all counts are fitted to the same models having the count name as the independent variable.

model
single character: "pois", "nb", "zinb", "zip", "all". If "all", all possible model are fitted.

level
Confidence level, default is 0.95.

...
Dots parameters are ignored.

Value

The list of fitted models. Names are names of original counts, an underline and a name of model used. confint is a matrix with the number of rows equal to the number of parameters. Rownames are names of parameters. The columns contain respectively lower and upper confidence intervals.

Examples

df <- data.frame(poisson = rpois(25, 0.3), binomial = rbinom(25, 1, 0.8))
fit_counts(df, model = "pois") 

#> $poisson_pois
#> $poisson_pois$coefficients
#> lambda 
#>   0.56 
#> 
#> $poisson_pois$confint
#>            lower     upper
#> lambda 0.3316685 0.9455224
#> 
#> $poisson_pois$BIC
#> [1] 52.99897
#> 
#> $poisson_pois$model
#> [1] "pois"
#> 
#> 
#> $binomial_pois
#> $binomial_pois$coefficients
#> lambda 
#>   0.88 
#> 
#> $binomial_pois$confint
#>            lower    upper
#> lambda 0.5794376 1.336468
#> 
#> $binomial_pois$BIC
#> [1] 52.84354
#> 
#> $binomial_pois$model
#> [1] "pois"
#> 
#>