Select the most appropriate model

Select the most appropriate model

select_model(fitlist)

Arguments

fitlist
a list of fits, as created by fit_counts.

Value

a data.frame with two columns: count representing the name of the count and chosen model with the model with the lowest BIC.

Examples

set.seed(1)
df <- data.frame(poisson1 = rpois(50, 2), 
                 poisson2 = rpois(50, 5),
                 zip1 = rZIP(50, 2, 0.7),
                 zip2 = rZIP(50, 5, 0.7))
fitlist_separate <- fit_counts(df, model = c("pois", "zip")) 
select_model(fitlist_separate)

#>      count chosen_model
#> 1 poisson1      Poisson
#> 2 poisson2      Poisson
#> 3     zip1          ZIP
#> 4     zip2          ZIP