Summary of estimates

Counts are fitted to model(s) using the count name as the explanatory variable. Estimates are presented in the table below along with the BIC values of their models. Estimated coefficients of models (lambda for all distributions, theta for NB and ZINB, r for ZIP and ZINB).

summary_fitlist(fitlist)

Arguments

fitlist
List is created by fitlist = fit_counts(count_list, model = "all"). For each count it creates distribution model

Value

Data frame with summarised results of all distribution models.

Countthe name of the original count.

lambda\(\lambda\) - Poisson mean, lower and upper confidence intervals.

BICBayesian information criterion

theta\(\theta\) - dispersion parameter

rprobability of excess zeros.

Examples

df <- data.frame(poisson = rpois(25, 0.3), binomial = rbinom(25, 1, 0.8))
fc <- fit_counts(df, model = "all")

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

summary_fitlist(fc)

#>      count    lambda     lower     upper      BIC     theta            r
#> 1  poisson 0.3600000 0.1873156 0.6918806 39.60860        NA           NA
#> 2 binomial 0.9200000 0.6113645 1.3844442 53.05443        NA           NA
#> 3  poisson 0.3600092 0.1873123 0.6919283 42.82758        NA 2.474199e-05
#> 4 binomial 0.9199837 0.6113507 1.3844264 56.27370        NA 9.785633e-06
#> 5  poisson 0.3600000 0.1873113 0.6918961 42.82776  11197.13           NA
#> 6 binomial 0.9200000 0.6113637 1.3844460 56.27341 198632.41           NA
#> 7  poisson 0.3600066 0.1873134 0.6919139 46.04647  51456.26 1.347321e-05
#> 8 binomial 0.9200015 0.6113651 1.3844473 59.49225 328028.61 1.831004e-07
#>     model
#> 1 Poisson
#> 2 Poisson
#> 3     ZIP
#> 4     ZIP
#> 5      NB
#> 6      NB
#> 7    ZINB
#> 8    ZINB