Summary of estimates

Counts are fitted to model(s) using the count name as the explanatory variable. Estimates are presented in the table below along with the BIC values of their models. Estimated coefficients of models (lambda for all distributions, theta for NB and ZINB, r for ZIP and ZINB).

summary_fitlist(fitlist)

Arguments

fitlist
a list of fits, as created by fit_counts.

Value

Data frame with summarised results of all distribution models.

Count: the name of the original count.

lambda: \(\lambda\) - Poisson mean, lower and upper confidence intervals.

BIC: Bayesian information criterion

theta: \(\theta\) - dispersion parameter

r: probability of excess zeros.

See also

Examples

df <- data.frame(poisson = rpois(25, 0.3), binomial = rbinom(25, 1, 0.8))
fc <- fit_counts(df, model = "all") 

#> Warning: iteration limit reached

#> Warning: iteration limit reached

summary_fitlist(fc) 

#>      count    lambda     lower     upper      BIC        theta            r
#> 1  poisson 0.3600000 0.1873133 0.6918888 43.76748           NA           NA
#> 2 binomial 0.7600000 0.4847687 1.1914961 51.64748           NA           NA
#> 3  poisson 0.8742175 0.3004691 2.5435433 44.58222           NA 5.882031e-01
#> 4 binomial 0.7600061 0.4847729 1.1915048 54.86645           NA 3.857156e-06
#> 5  poisson 0.3600000 0.1601595 0.8091934 45.65342 6.705164e-01           NA
#> 6 binomial 0.7600000 0.4847672 1.1914996 54.86660 5.730014e+04           NA
#> 7  poisson 0.8742013 0.3004503 2.5436082 47.80114 3.705731e+04 5.881948e-01
#> 8 binomial 0.7599975 0.4847659 1.1914951 58.08537 1.927820e+05 2.965500e-06
#>     model
#> 1 Poisson
#> 2 Poisson
#> 3     ZIP
#> 4     ZIP
#> 5      NB
#> 6      NB
#> 7    ZINB
#> 8    ZINB